requirements.txt
fastapi
uvicorn
jinja2
authlib
httpx
stripe
sqlalchemy
databases
python-dotenv
